tests/wpt/web-platform-tests/css/motion/parsing/offset-anchor-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Motion Path Module Level 1: getComputedStyle().offsetAnchor</title>
<link rel="help" href="https://drafts.fxtf.org/motion-1/#offset-anchor-property">
<meta name="assert" content="offset-anchor is 'auto' or a computed position.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<style>
  #target {
    font-size: 40px;
  }
</style>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("offset-anchor", "auto");

test_computed_value("offset-anchor", "left bottom", "0% 100%");
test_computed_value("offset-anchor", "center center", "50% 50%");
test_computed_value("offset-anchor", "right center", "100% 50%");
test_computed_value("offset-anchor", "center top", "50% 0%");
test_computed_value("offset-anchor", "center bottom", "50% 100%");
test_computed_value("offset-anchor", "calc(20% - 5em) center", "calc(20% - 200px) 50%");
test_computed_value("offset-anchor", "right 4em", "100% 160px");
test_computed_value("offset-anchor", "10px 20%");
test_computed_value("offset-anchor", "left -10px top -20%", "-10px -20%");
test_computed_value("offset-anchor", "right 10% bottom 20em", "90% calc(100% - 800px)");
test_computed_value("offset-anchor", "top 10px right -20%", "120% 10px");
test_computed_value("offset-anchor", "left 10% bottom 20em", "10% calc(100% - 800px)");
test_computed_value('offset-anchor', 'calc(10px - 0.5em) calc(10px + 0.5em)', '-10px 30px');
</script>
</body>
</html>
